# openapi.yaml
openapi: 3.0.3
info:
  title: UPISAS
  description: |-
    Unified Python interface for self-adaptive system exemplars.
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Some useful link here
  url: http://swagger.io
servers:
  - url: https://petstore3.swagger.io/api/v3
tags:
  - name: adaptation_options
    description: Get adaptation options
  - name: monitor
    description: Get data
  - name: execute
    description: Request a runtime adaptation
  - name: adaptation_options_schema
    description: Request the schemar of adaptation_options
  - name: monitor_schema
    description: Request the schema of monitoring
  - name: execute_schema
    description: Request the schema of execution

paths:
  /adaptation_options:
    get:
      tags:
        - adaptation_options
      summary: Get adaptation options from exemplar
      description: Used at the beginning of an exemplar run or whenever the adaptations options are changed during a run of the exemplar
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdaptationOptions'
        '400':
          description: Invalid status value
  /monitor:
    get:
      tags:
        - monitor
      summary: Get data from exemplar
      description: Used for runtime monitoring
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monitor'
        '400':
          description: Invalid status value
  /execute:
    put:
      tags:
        - execute
      summary: Enact a change
      description: Used to adapt the exemplar at runtime
      responses:
        '200':
          description: Successful operation
        '405':
          description: Invalid input
      requestBody:
        description: Apply an adaptation
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Execution'
  /adaptation_options_schema:
    get:
      tags:
        - adaptation_options_schema
      summary: Get the adaptation options schema
      description: Get the scheme of the adaptation options
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
  /monitor_schema:
    get:
      tags:
        - monitor_schema
      summary: Get the monitor schema
      description: Get the scheme of the monitored data
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
  /execute_schema:
    get:
      tags:
        - execute_schema
      summary: Get the execute schema
      description: Get the scheme of the execute data
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
components:
  schemas:
    AdaptationOptions:
      type: object
    Monitor:
      type: object
    Execution:
      type: object